pyproject.toml
[project]
name = "riedfm-g"
version = "0.1.0"
description = "Riemannian Discrete Flow Matching on Graphs - A Deep Graph Generative Foundation Model"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Yongli Mou"},
]
dependencies = [
    "torch>=2.6.0",
    "torch_geometric>=2.6.0",
    "geoopt>=0.5.0",
    "torchdiffeq>=0.2.4",
    "transformers>=4.48.0",
    "datasets>=3.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.19.0",
    "einops>=0.8.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.6.0",
    "networkx>=3.4",
    "pandas>=2.2.0",
    "tqdm>=4.67.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
riedfm-pretrain = "riedfm.cli.pretrain:main"
riedfm-finetune = "riedfm.cli.finetune:main"
riedfm-evaluate = "riedfm.cli.evaluate:main"
riedfm-generate = "riedfm.cli.generate:main"
riedfm-preprocess = "riedfm.cli.preprocess:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/riedfm"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B905",   # zip-without-strict (too noisy for torch code)
    "N806",   # variable in function should be lowercase (math convention: N, B, K, D)
    "N812",   # lowercase imported as non-lowercase (common in torch: F = nn.functional)
]

[tool.ruff.lint.isort]
known-first-party = ["riedfm"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = false
exclude = ["tests/", "scripts/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
